MAS-246 extract ProtectLevelCalculator
